Add layout support to the engine
This adds support for layouts to the engine. A layout is a template that
wraps another template. For example, a layout might be a header and footer
that wraps a template. The layout is registered with the engine and then
used in a template with the `layout` helper.

e.g.

```go
engine.Register("layouts/application", `{{partial "header"}} {{ChildContent}} {{partial "footer"}}`)
engine.Register("users/show", `{{layout "layouts/application"}}{{Name}}`)

engine.Render("users/show", map[string]any{"Name": "Fox Mulder"})
```